adding matching and gt_indices in reading
