+++ lib/writers/hpnet_dataset_writer.py
@@ -24,7 +24,6 @@ class HPNetDatasetWriter(BaseDatasetWriter):
         super().__init__(parameters)
 
     def step(self, points, normals=None, labels=None, features_data=[], noisy_points=None, filename=None, features_point_indices=None, **kwargs):
-
         if filename is None:
             filename = str(uuid.uuid4())
         
@@ -61,6 +60,10 @@ class HPNetDatasetWriter(BaseDatasetWriter):
                 
             h5_file.create_dataset('points', data=points)
             h5_file.create_dataset('normals', data=normals)
+            if 'gt_indices' in kwargs:
+                h5_file.create_dataset('gt_indices', data=kwargs['gt_indices'])
+            if 'matching' in kwargs:
+                h5_file.create_dataset('matching', data=kwargs['matching'])
 
             if labels is not None:
                 primitive_params = np.zeros((len(labels), 22))
